overview page created and similar style as profile
