+++ src/main/java/org/example/crimearchive/controllers/ProfileController.java
@@ -1,0 +1,44 @@
+package org.example.crimearchive.controllers;
+
+import org.example.crimearchive.cases.CaseService;
+import org.example.crimearchive.cases.Cases;
+import org.example.crimearchive.polis.Account;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import java.util.List;
+
+@Controller
+public class ProfileController {
+
+    CaseService caseService;
+
+    public ProfileController(CaseService caseService) {
+        this.caseService = caseService;
+    }
+
+
+    @GetMapping("/profile")
+    public String profilePage(@AuthenticationPrincipal Account user, Model model) {
+        List<Cases> caseList = caseService.getAuthzCases(user.getId());
+        model.addAttribute("currentUser", user);
+        model.addAttribute("cases", caseService.getReportsWithCaseNumber(caseList));
+        model.addAttribute("user", user);
+        return "profile";
+    }
+
+    @GetMapping("/caseoverview")
+    public String caseoverview(@AuthenticationPrincipal Account user,
+                               Model model, @RequestParam String casenumber) {
+        
+        var overviewList = caseService.getReportSet(casenumber);
+        System.out.println("amount of reports: " + overviewList.size());
+        model.addAttribute("currentUser", user);
+        model.addAttribute("rawcasenumber", casenumber);
+        model.addAttribute("reportList", overviewList);
+        return "caseoverview";
+    }
+}
