+++ src/main/resources/templates/caseoverview.html
@@ -99,6 +99,14 @@
                             </div>
                             <div class="file-status">1.8 MB · Klar för uppladdning</div>
                         </div>
+                        <table>
+                            <th:block th:each="report : ${reportList}">
+                                <tr>
+                                    <th class="file-name" th:text="${report.getEvent()}"></th>
+                                    <td class="file-meta">Rapport beskrivning</td>
+                                </tr>
+                            </th:block>
+                        </table>
 
                         <div class="evidence-row">
                             <div class="file-meta">
